test_aggregate is the wrong place to put
